requirements.txt
Django==2.2.7
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.10.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
